requirements.txt
streamlit
opencv-python-headless
numpy
requests
streamlit-image-comparison
Pillow